+++ public/static/api-dnb.json
@@ -1,0 +1,6 @@
+{
+  "0xE94E4f16ad40ADc90C29Dc85b42F1213E034947C": "Bank of China (DnB)",
+  "0xa61B056dA0084a5f391EC137583073096880C2e3": "DBS (DnB)",
+  "0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e": "China Oil (DnB)",
+  "0x4BD9B0838FcC487DBc883ad0E417aEd98001268a": "OCBC (DnB)"
+}
