+++ src/components/AddressResolver/AddressResolved.tsx
@@ -7,46 +7,44 @@ import { ThirdPartyAPIEntryProps, useThirdPartyAPIEndpoints } from "./../../comm
 export const AddressResolved = () => {
   const { addressBook } = useAddressBook();
   const { thirdPartyAPIEndpoints } = useThirdPartyAPIEndpoints();
-  const [addressResolved, setAddressResolved] = useState({ ...addressBook } as AddressBook); // load addressbook as base
-
-  const APIrequests: {}[] = [];
-  let APIAddresses = {};
-
-  const getThirdPartyAddresses = (url: string) => {
-    return axios
-      .get(url)
-      .then((response) => {
-        const lowerCaseAddressesObj = transform(response.data, (result: any, value: string, key: string) => {
-          result[key.toLowerCase()] = value;
-        });
-        APIAddresses = { ...lowerCaseAddressesObj, ...APIAddresses };
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  const prepAPIAddressEndpoints = ({ endpoint }: ThirdPartyAPIEntryProps) => {
-    APIrequests.push(getThirdPartyAddresses(endpoint));
-  };
+  const [addressResolved, setAddressResolved] = useState({} as AddressBook);
 
   useEffect(() => {
+    const APIrequests: {}[] = [];
+    let APIAddresses = {};
+
+    const getThirdPartyAddresses = (url: string) => {
+      return axios
+        .get(url)
+        .then((response) => {
+          const lowerCaseAddressesObj = transform(response.data, (result: any, value: string, key: string) => {
+            result[key.toLowerCase()] = value;
+          });
+          APIAddresses = { ...lowerCaseAddressesObj, ...APIAddresses };
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    };
+
+    const prepAPIAddressEndpoints = ({ endpoint }: ThirdPartyAPIEntryProps) => {
+      APIrequests.push(getThirdPartyAddresses(endpoint));
+    };
+
     if (thirdPartyAPIEndpoints.length > 0) {
       thirdPartyAPIEndpoints.forEach(prepAPIAddressEndpoints);
       axios.all(APIrequests).finally(() => {
-        setAddressResolved({ ...addressResolved, ...APIAddresses });
+        setAddressResolved({ ...addressBook, ...APIAddresses });
       });
     } else {
       setAddressResolved({ ...addressBook });
     }
-  }, [thirdPartyAPIEndpoints]);
-
-  console.log({}, "AddressResolved");
+  }, [addressBook, thirdPartyAPIEndpoints]);
 
   return (
     <div className="row my-4">
       <div className="col-12">
-        <b>Address Resolved (AssetInformationPanel):</b>
+        <b>AddressResolved (AssetInformationPanel):</b>
         {isEmpty(addressResolved) ? (
           <p>Nothing from local addressbook or third party api.</p>
         ) : (
